requirements.txt
google-cloud-storage
google-cloud-bigquery
google-api-python-client
python-dotenv
pandas
pyarrow
